+++ Core/GameEngine/Source/GameClient/GUI/Gadget/GadgetVerticalSlider.cpp
@@ -375,7 +375,7 @@ WindowMsgHandledType GadgetVerticalSliderSystem( GameWindow *window, UnsignedInt
 		// ------------------------------------------------------------------------
 		case GSM_SET_SLIDER:
 		{
-			Int newPos = (Int)mData1;
+			Int newPos = (Int)(intptr_t)mData1;
 			GameWindow *child = window->winGetChild();
 
 			// TheSuperHackers @fix No longer reject out of bounds positions to prevent
@@ -402,8 +402,8 @@ WindowMsgHandledType GadgetVerticalSliderSystem( GameWindow *window, UnsignedInt
 
 			window->winGetSize( &size.x, &size.y );
 
-			s->minVal = (Int)mData1;
-			s->maxVal = (Int)mData2;
+			s->minVal = (Int)(intptr_t)mData1;
+			s->maxVal = (Int)(intptr_t)mData2;
 			s->numTicks = (Real)( size.y-GADGET_SIZE)/(Real)(s->maxVal - s->minVal);
 			s->position = s->minVal;
 
@@ -448,8 +448,8 @@ WindowMsgHandledType GadgetVerticalSliderSystem( GameWindow *window, UnsignedInt
 		// ------------------------------------------------------------------------
 		case GGM_RESIZED:
 		{
-			Int width = (Int)mData1;
-//			Int height = (Int)mData2;
+			Int width = (Int)(intptr_t)mData1;
+//			Int height = (Int)(intptr_t)mData2;
 			GameWindow *thumb = window->winGetChild();
 
 			if( thumb )
